rag_model/Agno_RAG_v01/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
pydantic>=2.0.0
pydantic-settings>=2.0.0
supabase>=2.0.0
supabase-py[asyncio]>=2.0.0  # For async support
python-dotenv>=1.0.0
langchain>=0.1.0
langchain-community>=0.0.10
google-generativeai>=0.3.0
tiktoken>=0.5.0
numpy>=1.24.0
httpx>=0.24.0
jinja2>=3.1.0
agno>=0.1.0